refactor in rest. RestServiceBase is added. StreamsSourceRestService and BroadcastRestService extend it.
