+++ src/main/java/io/antmedia/rest/BroadcastRestService.java
@@ -86,7 +86,7 @@ import io.antmedia.webrtc.api.IWebRTCAdaptor;
 )
 @Component
 @Path("/")
-public class BroadcastRestService {
+public class BroadcastRestService extends RestServiceBase{
 
 	public static class BroadcastStatistics {
 
@@ -122,20 +122,7 @@ public class BroadcastRestService {
 	public static final int MP4_DISABLE = -1;
 	public static final int MP4_NO_SET = 0;
 
-	@Context
-	private ServletContext servletContext;
-
-	private IScope scope;
-
-
-	private ApplicationContext appCtx;
-
-	private AntMediaApplicationAdapter app;
-
-	private IDataStore dataStore;
-
 	private AppSettings appSettings;
-	private DataStoreFactory dataStoreFactory;
 
 	public interface ProcessBuilderFactory {
 		Process make(String...args);
@@ -446,43 +433,14 @@ public class BroadcastRestService {
 	public Result addSocialEndpoint(@ApiParam(value = "Stream id", required = true) @FormParam("id") String id,
 									@ApiParam(value = "name of the service like facebook, youtube, periscope in order to have successfull operation. Social network must be authorized in advance", required = true)
 									@FormParam("serviceName") String endpointServiceId) {
-		boolean success = false;
-		String message = null;
 		Broadcast broadcast = lookupBroadcast(id);
-
-		if (broadcast != null) {
-			Map<String, VideoServiceEndpoint> endPointServiceList = getEndpointList();
-
-			if (endPointServiceList != null) {
-
-				VideoServiceEndpoint videoServiceEndpoint = endPointServiceList.get(endpointServiceId);
-
-				if (videoServiceEndpoint != null) {
-					Endpoint endpoint;
-					try {
-						endpoint = videoServiceEndpoint.createBroadcast(broadcast.getName(),
-								broadcast.getDescription(), id, broadcast.isIs360(), broadcast.isPublicStream(),
-								720, true);
-						success = getDataStore().addEndpoint(id, endpoint);
-
-					} catch (Exception e) {
-						logger.error(ExceptionUtils.getStackTrace(e));
-						message = e.getMessage();
-					}
-				}
-				else {
-					message = endpointServiceId + " endpoint does not exist in this app.";
-					logger.warn(message);
-				}
-			} else {
-				message = "No social endpoint is defined for this app. Consult your app developer";
-				logger.warn(message);
-			}
-		} else {
-			message = "No broadcast exist with the id specified";
-			logger.warn(message);
+		
+		boolean success = addSocialEndpoints(broadcast, endpointServiceId);
+		String message = "";
+		if(!success) {
+			message  = endpointServiceId+" endpoint can not be added to "+broadcast.getStreamId();
 		}
-
+		
 		return new Result(success, message);
 	}
 
@@ -1701,52 +1659,8 @@ public class BroadcastRestService {
 		return getDataStore().getBroadcastCount();
 	}
 
-	protected Map<String, VideoServiceEndpoint> getEndpointList() {
-		return getApplication().getVideoServiceEndpoints();
-	}
-
 	protected List<VideoServiceEndpoint> getEndpointsHavingErrorList(){
 		return getApplication().getVideoServiceEndpointsHavingError();
-	}
-
-	@Nullable
-	private ApplicationContext getAppContext() 
-	{
-		if (appCtx == null && servletContext != null) {
-			appCtx = (ApplicationContext) servletContext
-					.getAttribute(WebApplicationContext.ROOT_WEB_APPLICATION_CONTEXT_ATTRIBUTE);
-
-		}
-		return appCtx;
-	}
-
-	public IDataStore getDataStore() {
-		if (dataStore == null) {
-			dataStore = getDataStoreFactory().getDataStore();
-		}
-		return dataStore;
-	}
-
-	public void setDataStore(IDataStore dataStore) {
-		this.dataStore = dataStore;
-	}
-
-	public AntMediaApplicationAdapter getApplication() {
-		if (app == null) {
-			ApplicationContext appContext = getAppContext();
-			if (appContext != null) {
-				app = (AntMediaApplicationAdapter) appContext.getBean(AntMediaApplicationAdapter.BEAN_NAME);
-			}
-		}
-		return app;
-	}
-
-	/**
-	 * this is for testing
-	 * @param app
-	 */
-	public void setApplication(AntMediaApplicationAdapter app) {
-		this.app = app;
 	}
 
 	private AppSettings getAppSettings() {
@@ -1759,36 +1673,8 @@ public class BroadcastRestService {
 		return appSettings;
 	}
 
-	public IScope getScope() {
-		if (scope == null) {
-			scope = getApplication().getScope();
-		}
-		return scope;
-	}
-
-	public void setScope(IScope scope) {
-		this.scope = scope;
-	}
-
-	public void setAppCtx(ApplicationContext appCtx) {
-		this.appCtx = appCtx;
-	}
-
 	public void setAppSettings(AppSettings appSettings) {
 		this.appSettings = appSettings;
-	}
-
-	public DataStoreFactory getDataStoreFactory() {
-		if(dataStoreFactory == null) {
-			WebApplicationContext ctxt = WebApplicationContextUtils.getWebApplicationContext(servletContext); 
-			dataStoreFactory = (DataStoreFactory) ctxt.getBean(IDataStoreFactory.BEAN_NAME);
-		}
-		return dataStoreFactory;
-	}
-
-
-	public void setDataStoreFactory(DataStoreFactory dataStoreFactory) {
-		this.dataStoreFactory = dataStoreFactory;
 	}
 
 	public ProcessBuilderFactory getProcessBuilderFactory() {
